requirements.txt
pandas
python-dotenv
requests
google-cloud-bigquery
pyarrow
pandas_gbq